Update Connection.groovy
We had an issue with our Sentry server (separate server from the clustered app server) was not accessible over our network. When this issue occurred, and many errors were occurring on the app server, the number of HTTPURLConnection's available in the "send" method became limited and the network connect timeout time, quite being long, caused the app servers to eventually  lockup completely.

Using shorter timeouts and doing the actual work in "send" asynchronously seemed like a reasonable fix. We have tested this solution and are satisfied with the results. Hope this fix may be of use to someone else and/or an alternate/better solution can be suggested.